watcher: Move methods accessing pods to a separate file pod.go
There are no functional changes in this commit, only moving code around. It's a
preparation to further refactoring of the k8s watcher that will decouple the
core k8s watcher and resource-specific functionalities, and allow for reusing
of informer factories. Also removed unused global constants.
